requirements.txt
dj-database-url==0.4.1
Django==1.11.3
gunicorn==19.7.1
psycopg2==2.6.2
whitenoise==3.2
pillow==4.2.1